+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Anticipated release: July 9, 2021
 
 #### 🚀 New features
 
+- Allow switching between states/affiliations when applicable ([#2581])
 - Update the delete process for FFY ([#2996])
 - updated Assurance and Compliance text on APD review and export screen to show more detailed information. ([#3169]))
 
@@ -17,11 +18,13 @@ Anticipated release: July 9, 2021
 
 - Created endpoint to allow a user to switch states ([#3030])
 - Added security check to state/:id endpoint so only users with access to that state can get the data ([#3129])
+- endpoint for pulling a user's affiliations ([#3207])
 
 # Previous releases
 
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
+[#2581]: https://github.com/CMSgov/eAPD/issues/2581
 [#2996]: https://github.com/CMSgov/eAPD/issues/2996
 [#3030]: https://github.com/CMSgov/eAPD/issues/3030
 [#3180]: https://github.com/CMSgov/eAPD/issues/3180
@@ -29,3 +32,5 @@ See our [release history](https://github.com/CMSgov/eAPD/releases)
 [#3169]: https://github.com/CMSgov/eAPD/issues/3169
 [#3238]: https://github.com/CMSgov/eAPD/issues/3238
 [#3194]: https://github.com/CMSgov/eAPD/issues/3194
+
+[#3207]: https://github.com/CMSgov/eAPD/issues/3207